requirements.txt
mss
pillow
imagehash
pyautogui
openai
openpyxl
paddlepaddle
paddleocr
pywin32
pyinstaller
